python/tracing/vertexai-openinference/requirements.txt
google-cloud-aiplatform>=1.49.0
openinference-instrumentation-vertexai
respan-tracing
respan-instrumentation-openinference
python-dotenv
requests
